requirements.txt
python-decouple
telethon
youtube-dl
